requirements.txt
torch
torchvision
numpy
matplotlib
argparse
Pillow
